최종 배포 버전 merge

+++ main_project/src/main/java/NutrientsCoders/main_project/food/dto/FoodPostDto.java
@@ -8,7 +8,7 @@ public class FoodPostDto {
   
   private Long servingSize;
   
-  private Long kcal;
+  private Double kcal;
   
   private Double carbo;
